schemas: remove RecordSchema [+]
- It is up to individual services to define their own RecordSchema.
- Example of this is done in the tests where the prvious RecordSchema(s)
  were moved.
