+++ invenio_records_resources/services/records/components.py
@@ -86,15 +86,14 @@ class FilesOptionsComponent(ServiceComponent):
 
     def _validate_files_enabled(self, record, enabled):
         """Validate files enabled."""
-        if not enabled:
-            if record.files.values():
-                raise ValidationError(
-                    _("You must first delete all files to set the record to "
-                      "be metadata-only."),
-                    field_name="files.enabled"
-                )
+        if not enabled and record.files.values():
+            raise ValidationError(
+                _("You must first delete all files to set the record to "
+                  "be metadata-only."),
+                field_name="files.enabled"
+            )
 
-    def assign_files_enabled(self, identity, enabled, record=None, **kwargs):
+    def assign_files_enabled(self, enabled, record=None, **kwargs):
         """Assign files enabled.
 
         This is a public interface so that it can be reused elsewhere
@@ -107,10 +106,10 @@ class FilesOptionsComponent(ServiceComponent):
         """Inject parsed files options in the record."""
         # presence is guaranteed by schema
         enabled = data["files"]["enabled"]
-        self.assign_files_enabled(identity, enabled, record, **kwargs)
+        record.files.enabled = enabled
 
     def update(self, identity, data=None, record=None, **kwargs):
         """Inject parsed files options in the record."""
         # presence is guaranteed by schema
         enabled = data["files"]["enabled"]
-        self.assign_files_enabled(identity, enabled, record, **kwargs)
+        self.assign_files_enabled(enabled, record, **kwargs)
