+++ tests/factories/test_service.py
@@ -1,8 +1,18 @@
+# -*- coding: utf-8 -*-
+#
+# This file is part of Invenio.
+# Copyright (C) 2020-2021 CERN.
+# Copyright (C) 2020-2021 Northwestern University.
+#
+# Invenio-Records-Resources is free software; you can redistribute it and/or
+# modify it under the terms of the MIT License; see LICENSE file for more
+# details.
+
 from invenio_records_permissions import RecordPermissionPolicy
 from invenio_records_permissions.generators import AnyUser
+from mock_module.schemas import RecordSchema
 
 from invenio_records_resources.factories.factory import RecordTypeFactory
-from invenio_records_resources.services.records.schema import RecordSchema
 
 
 def test_simple_flow(app, identity_simple, db):
